+++ silx/math/fit/__init__.py
