+++ silx/math/fit/curve_fit.py
